+++ integration-tests/src/test/scala/peloton/actors/FooActor.scala
@@ -27,8 +27,8 @@ object FooActor:
   given CanAsk[Get, GetResponse] = canAsk
     
   def spawn(name: String, persistenceId: PersistenceId)(using DurableStateStore)(using actorSystem: ActorSystem): IO[ActorRef[Message]] =
-    actorSystem.spawn(
-      name            = name,
+    actorSystem.spawnDurableStateActor(
+      name            = Some(name),
       persistenceId   = persistenceId, 
       initialState    = State(), 
       initialBehavior =